feat: Add a few tests and propType checking
Didn't add tests for most of the components, because
it kept showing the error for React10.useContext() returning
null, which seems to be indrectly telling me that <Link />
cannot be used outside of the react router. But the tests
are outside of the react router.

https://github.com/remix-run/react-router/issues/12989

So, I'll add tests for other components Later when I've found
the solutions for this problem